+++ app/src/main/kotlin/com/aliucord/manager/patcher/signing/KeySet.kt
@@ -1,0 +1,9 @@
+package com.aliucord.manager.patcher.signing
+
+import java.security.PrivateKey
+import java.security.cert.X509Certificate
+
+data class KeySet(
+    val publicKey: X509Certificate,
+    val privateKey: PrivateKey,
+)
